backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
yfinance>=1.1.0
pandas>=2.2.0
numpy>=1.26.4
scikit-learn>=1.5.0
transformers>=4.44.0
requests>=2.32.0
python-dateutil>=2.9.0
ta>=0.11.0
pydantic>=2.0.0
xgboost>=2.0.0
nltk>=3.8.1
google-generativeai>=0.3.0
# Note: torch is installed separately in Dockerfile to use CPU-only index
